+++ library/Dlayer/DesignerTool/ContentManager/Page/SubTool/Styling/Model.php
@@ -1,147 +1,83 @@
 <?php
 
 /**
- * Data model for the styling sub tool
+ * Model class for styling sub tool, no direct database access calls utility model class.
  *
  * @author Dean Blackborough <dean@g3d-development.com>
  * @copyright G3D Development Limited
  * @license https://github.com/Dlayer/dlayer/blob/master/LICENSE
  */
-class Dlayer_DesignerTool_ContentManager_Page_SubTool_Styling_Model extends Zend_Db_Table_Abstract
+class Dlayer_DesignerTool_ContentManager_Page_SubTool_Styling_Model
 {
-    /**
-     * Fetch the background color assigned to the content page
-     *
-     * @param integer $site_id
-     * @param integer $id
-     *
-     * @return string|false
-     */
-    public function backgroundColor($site_id, $id)
+    private $model;
+
+    public function __construct()
     {
-        $sql = "SELECT background_color 
-                FROM user_site_page_styling_page_background_color
-                WHERE site_id = :site_id 
-                AND page_id = :page_id 
-                LIMIT 1";
-        $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':site_id', $site_id);
-        $stmt->bindValue(':page_id', $id);
-        $stmt->execute();
-
-        $result = $stmt->fetch();
-
-        if ($result !== false) {
-            return $result['background_color'];
-        } else {
-            return false;
-        }
+        $this->model = new Dlayer_Model_DesignerTool_ContentManager_PageStyling();
     }
 
     /**
-     * Check to see if there is an existing background colour defined for the page
+     * Fetch the background colour assigned to the content section of the page
      *
      * @param integer $site_id
-     * @param integer $id
+     * @param integer $page_id
+     *
+     * @return array|false
+     */
+    public function contentBackgroundColor($site_id, $page_id)
+    {
+        return $this->model->getContentAttributeValue($site_id, $page_id, 'background-color');
+    }
+
+    /**
+     * Fetch the background colour assigned to the html
+     *
+     * @param integer $site_id
+     *
+     * @return array|false
+     */
+    public function htmlBackgroundColor($site_id)
+    {
+        return $this->model->getHtmlAttributeValue($site_id, 'background-color');
+    }
+
+    /**
+     * Check to see if there is an existing background colour defined for the content container
+     *
+     * @param integer $site_id
+     * @param integer $page_id
      *
      * @return integer|false
      */
-    public function existingBackgroundColor($site_id, $id)
+    public function existingContentBackgroundColorId($site_id, $page_id)
     {
-        $sql = "SELECT id 
-                FROM user_site_page_styling_page_background_color 
-                WHERE site_id = :site_id 
-                AND page_id = :page_id 
-                LIMIT 1";
-        $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':site_id', $site_id);
-        $stmt->bindValue(':page_id', $id);
-        $stmt->execute();
-
-        $result = $stmt->fetch();
-        if ($result !== false) {
-            return intval($result['id']);
-        } else {
-            return false;
-        }
+        return $this->model->existingContentAttributeId($site_id, $page_id, 'background-color');
     }
 
     /**
-     * Add a new background color to the page
+     * Add a new background color for the content container
      *
      * @param integer $site_id
-     * @param integer $id
-     * @param string $color_hex
+     * @param integer $page_id
+     * @param string $color
      *
      * @return boolean
      */
-    public function addBackgroundColor($site_id, $id, $color_hex)
+    public function addContentBackgroundColor($site_id, $page_id, $color)
     {
-        $sql = "INSERT INTO user_site_page_styling_page_background_color 
-                (site_id, page_id, background_color) 
-                VALUES 
-                (:site_id, :page_id, :background_color)";
-        $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':site_id', $site_id);
-        $stmt->bindValue(':page_id', $id);
-        $stmt->bindValue(':background_color', $color_hex);
-
-        return $stmt->execute();
+        return $this->model->addContentAttributeValue($site_id, $page_id, 'background-color', $color);
     }
 
     /**
-     * Edit the background color for the page, optionally delete any set values
+     * Edit the background color for the content page, optionally delete any existing value
      *
      * @param integer $id
-     * @param string $color_hex
+     * @param string $color
      *
      * @return boolean
      */
-    public function editBackgroundColor($id, $color_hex)
+    public function editContentBackgroundColor($id, $color)
     {
-        if (strlen($color_hex) !== 0) {
-            return $this->updateBackgroundColor($id, $color_hex);
-        } else {
-            return $this->deleteBackgroundColor($id);
-        }
-    }
-
-    /**
-     * Update the background colour for the row
-     *
-     * @param integer $id
-     * @param string $color_hex
-     *
-     * @return boolean
-     */
-    protected function updateBackgroundColor($id, $color_hex)
-    {
-        $sql = "UPDATE user_site_page_styling_page_background_color 
-                SET background_color = :background_color 
-                WHERE id = :id 
-                LIMIT 1";
-        $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':id', $id);
-        $stmt->bindValue(':background_color', $color_hex);
-
-        return $stmt->execute();
-    }
-
-    /**
-     * Remove a background colour for the selected row
-     *
-     * @param integer $id
-     *
-     * @return boolean
-     */
-    protected function deleteBackgroundColor($id)
-    {
-        $sql = "DELETE FROM user_site_page_styling_page_background_color 
-                WHERE id = :id 
-                LIMIT 1";
-        $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':id', $id);
-
-        return $stmt->execute();
+        return $this->model->editContentAttributeValue($id, $color);
     }
 }