+++ library/Dlayer/Model/View/Page/Styling.php
@@ -235,19 +235,25 @@ class Dlayer_Model_View_Page_Styling extends Zend_Db_Table_Abstract
      */
     private function pageBackgroundColors()
     {
-        $sql = "SELECT page_id, background_color 
-                FROM user_site_page_styling_page_background_color 
-                WHERE site_id = :site_id 
-                AND page_id = :page_id";
+        $sql = "SELECT 
+                    `page_id`, 
+                    `value` 
+                FROM 
+                    `user_site_page_styling` 
+                WHERE 
+                    `site_id` = :site_id AND 
+                    `page_id` = :page_id AND 
+                    `attribute` = :attribute";
         $stmt = $this->_db->prepare($sql);
-        $stmt->bindValue(':site_id', $this->site_id);
-        $stmt->bindValue(':page_id', $this->page_id);
+        $stmt->bindValue(':site_id', $this->site_id, PDO::PARAM_INT);
+        $stmt->bindValue(':page_id', $this->page_id, PDO::PARAM_INT);
+        $stmt->bindValue(':attribute', 'background-color', PDO::PARAM_STR);
         $stmt->execute();
 
         $styles = array();
 
         foreach ($stmt->fetchAll() as $row) {
-            $styles[intval($row['page_id'])] = $row['background_color'];
+            $styles[intval($row['page_id'])] = $row['value'];
         }
 
         return $styles;